Prometheus-Openstack: Allow -* suffix for named metrics ports
The name of a port needs to be unique in a pod,
but there might be more than container with metrics in a pod.
This extends the convention for ports named 'metrics' also
to ports with a suffix separated by hyphen ('metrics-foo'),
so one doesn't have to resort to a pod annotation.