[Ironic] Additional swift reseller role
We are migrating to CCloud conform role namings and introduced a new Swift reseller role:
`cloud_objectstore_admin`. To make the migration as smooth as possible, the new role will
be already assigned to the `service` user in the Default domain.
Removal off the old role assignment will happen later.