[sap-seeds] Hana-only BB regions don't need CUSTOM_MEMORY_RESERVABLE_MB
We introduced CUSTOM_MEMORY_RESERVABLE_MB as resource to guard against a
cluster not getting overloaded with memory reservation and thus not
being able to sustain more failed nodes than we have failover hosts.
This is fine if we run mixed workloads on a BB and the smaller VMs can
fill up the space not available to hana_* flavors.

With regions switching to hana-only BBs, this hinders spawning new
hana_* flavors and keeps those hana-only BBs less utilized than we want.
Therefore, we remove the requirement on CUSTOM_MEMORY_RESERVABLE_MB from
the hana_* flavors if we run with hana-only BBs.