swift: correctly represent daemonset selectors as they exist in the wild
We originally created DaemonSets in apiVersion: extensions/v1beta1
without explicit selectors. When upgrading to apps/v1, we added
selectors, but those selectors did not include these label matches
below.

Kubernetes had, however, chosen to include these label matches
when upgrading the internal representation of the existing DaemonSets
from extensions/v1beta1 to apps/v1. So we have these label matches in
the selectors in the regions that existed at that time. We then built up
new regions, so the selectors there do not have those label matches.

Since we now have more strict validation on Helm objects, we need to
have the selector in the Helm template accurately reflect the reality in
the Kubernetes API.

I'm only fixing the selectors on daemonsets that are sensitive to
restarts. On components where a full-blown recreate is feasible, I will
delete the daemonsets and deploy fresh versions instead.