[manila] Add ProxySQL side-cars
As this chart works with a pre-upgrade hook,
care has to be taken when using
proxysql.mode=unix_socket

The pre-upgrade hook works with the configmap using the old values,
while the chart would be rendered with the new values.
The only mode that works with both is
proxysql.mode=host_alias

So when either switching from no proxysql to unix_socket
or the other way around, a deployment in with the host_alias mode
is needed in between.