[openstack/utils] Use first dbname & user as default
All services where setting a dbUser and dbName in the current
chart. That works for a single instance, but not in general.
Instead use the logic that the first database listed
under mariadb.databases is the default database,
and that there is user with the same key (but not necessarily
the same name) as the database.