[openstack/utils] Add coordination section and pvc
Add a shared ini-section snippet and related snippets for volume
mounts, pvc, and volumes which may be activated
by setting coordinationBackend=file

The tooz library offers various drivers for the coordination of
the openstack processes. We started out with memcached, as it
presumably fast and already part of the stack.
The major downside of memcached is not offering any persistence,
so should memcached restart any locks would be lost, possibly
resulting in two processes entering a section of code,
which should only run exclusively in one process.

Since we are now regularly updating the OS and services,
all services including databases and memcached are
getting restarted much more often.

The mysql/mariadb/postgresql drivers are using session locks,
which will also be lost during a restart of the database.
We opted for the "file" driver using a shared filesystem (NFS)
at least intermediary solution to get locks, which are restart safe.
To keep the option open of different backends, this change introduces
the config-value coordinationBackend, so we can eventually easily
switch to a different one should the need arise.