[Cinder] Disable incremental backups
This patch adds the ability to set the cinder option
to disable incremental backups.  The patch defaults
the option to On.

We are disabling incremental backups because the
incremental backups with vmware volumes doesn't work
properly because the volumes are compressed and the
sizes of each blocks change depending on the uncompressed data.
We can't fix this in an easy way, so disabling incremental
backups is the only way so far.