common/postgresql: remove pgbouncer
This part has not been used in a long time (ever since the OpenStack
core services moved to MariaDB), so it probably will have bit-rot
already and it's best to get rid of it. (If you disagree, you can
restore the pgbouncer part from the Git history.)

I checked that the following charts produce an empty `helm diff` with
this change: arc, castellum, clair, elektra, keppel, limes, lyra,
sentry, tenso.