Nova: Add a secret per cell
With the Xena upgrade, we lose the API endpoint os-cells as cells v1 are
getting removed. We had previously extended that endpoint to let the
vcenter-operator retrieve connection information per cell and expose
these to its templates for configured nova-compute-bb* into the right
cell.

Our new approach is to expose the same information as k8s Secret per
cell, identified by the appropriate labels. The information is basically
the same, only lacking information about the UUID of the cell, as that's
a Nova-runtime-internal and not a configurable.