[nova] Move to track_instance_changes setting
The old DEFAULT/scheduler_tracks_instance_changes setting was renamed to
filter_scheduler/track_instance_changes in Ocata. Xena removes reading
the old name for setting that setting. Therefore, we rename all
occurrences to the new setting in preparation for the Xena upgrade.
Given the rename happening in Ocata, this is compatible with both Xena
and Rocky.