[nova] Remove glance/api_servers setting
Using this setting was deprecated starting Queens where [0] was
implemented. Instead of manually specifying api_servers for glance,
standard keystoneauth1 endpoint discovery takes place - in Nova using
the internal endpoint as a priority. Since we previously had set the
internal endpoint in api_servers, there's no need to set the equivalent
in keystoneauth1 option endpoint_override.

We don't need to set any other settings as the other settings are used
for authentication and authentication is handled by using the user's
token only.

Given the change to keystoneauth1 was implemented in Queens, this change
is compatible with Rocky and Xena.

[0] https://specs.openstack.org/openstack/nova-specs/specs/queens/implemented/use-ksa-adapter-for-endpoints.html