[nova] Support dhcp_domain setting in Xena
When upgrading to Xena, the previously deprecated DEFAULT/dhcp_domain
setting will move to api/dhcp_domain. Since this move happens in Train,
we cannot just move the setting. Instead, we rely on checking the
image-version to see if we're on Rocky or Xena and add the setting to
the appropriate section.

Using the image-version for checking the release is already done in the
_db-online-migrate.tpl.

We could also set the setting in both places, as it should be ignored
under [api] in Rocky and under [DEFAULT] in Xena. It feels better to
have a clean config in production though, that's why I opted for setting
the setting conditionally.