[nova] Remove consoleauth support
Starting with Rocky, Nova doesn't use the consoleauth service anymore
unless explicitly enabled in the [workarounds] section. Instead, tokens
are saved in the DB and read from there. We only needed consoleauth
enabled during the migration to Rocky and have disabled it since.

Since we're updating to Xena, we now remove the consoleauth Deployment
altogether, because there the consoleauth service doesn't exist in Nova
anymore.