[nova] Rework api-paste.ini
This is a general update of the api-paste.ini to make it more in line
with upstream. There are a couple of changes we can already do for Rocky
and more changes we need to do for Xena, guarded by some conditionals.

* Removal of "noauth2"
Usage of "noauth2" is deprecated in Xena and we don't use it in Rocky
already, because it's only meant for development and circumvents any
authentication mechanism i.e. keystone. We don't need that so we remove
it.

* Move "healthcheck" from filter to app
oslo.middleware states already since at least Rocky, that healthcheck
should not be run as a filter anymore, but as an app instead [0]. This
also means that we remove "healtcheck" from any pipeline definition.

* Split API versions endpoints for Xena
While Rocky doesn't support that (the feature comes in with Ussuri),
there are different versions endpoints for the legacy (/v2) and the
current (/v2.1) API endpoints. Since this only works in Xena, we guard
the splitting by conditionals.

* Re-order of pipelines' filters
Upstream had a different order of filters namely for "faultwrap" and
"sizelimit" which we now also adhere to. This moves "osprofiler" to after
those filters and thus doesn't allow profiling them.

* Add "request_log" filter
While we're still running with the eventlet WSGI server on both Rocky
and Xena, we will eventually have to migrate away from it. To still get
the same logging of requests, upstream implemented a "request_log"
filter, which will only log if Nova doesn't run on eventlet. We
integrate this filter into our pipelines to prepare for the switch
already and to be more in line with upstream.

* Removal of "request_id" and "legacy_ratelimit" filters
The filters were unused.

* Removal of "osapi_compute_app_legacy_v2" app
The app is no longer in use, as it was deprecated and is replaced with
"openstack_compute_api_v21_legacy_v2_compatible".

[0] https://github.com/openstack/oslo.middleware/blob/stable/rocky/oslo_middleware/healthcheck/__init__.py#L408-L410