[nova] Change API unexpected exception message
With Xena, we're easily able to configure the "API unexpected exception"
message by setting the "support" config option in the "release" file [0]. We
do this here and point to the "Contact Us" site of our documentation
instead of upstream's bug tracker.

We only have to do this for the nova-api Deployment, because that's the
only user-facing site. nova-api-metadata is called through Neutron and
shouldn't forward the messages of Nova directly.

[0] https://github.com/sapcc/nova/blob/b0efa22cd37073d49519ddff02bdb69e5df4d5a7/nova/version.py#L62-L63