gatekeeper: add a unit test for the "add_support_labels" lib module
This is a proposal on how to avoid duplication of work in the test suite
by having separate "unit test" policies for the individual lib modules.
By exercising the different code paths inside the lib module within the
unit test, we can avoid having lots of similar test cases in the
end-to-end tests for the actual policies.