[nova] Policy: Add doc comments to each policy
Unknown policies have a preliminary "# ??" added to make them easily
identifiable. These policies will either be removed in a subsequent
commit or have their comment filled in.

Additional policies mentioned in the docs but which weren't in the
config file, are added at the bottom with only the comment section.
These should be adapted to the rest of the rules.