[placement] Add policy.yaml for Yoga
While we run placement without policy overrides in Rocky, we have to
add some overrides to keep things working when upgrading to Yoga. The
reason for that is, that Rocky's default policy is to allow any
"role:admin" user to access Placement, while in Yoga only admins with
system-scoped tokens can access the important parts.

We change this default, to include our cloud_compute_admin role for
admin APIs, too.

It should be fine to roll this file out with Rocky, too, because
oslo.policy and placement should not know about a policy.yaml in that
version.

The new defaults include a reader mode and users being able to get a
usages overview of their own project to accommodate for Nova switching
the Quota to counting via Placement in Yoga.