[Cinder] Add new SAPPoolDownFilter to default filters
This patch moves the CapacityFilter to the last filter to run as well
as adding a new SAPPoolDownFilter, which will filter out all pools
that are not marked as 'up' by the vmware driver.

The CapacityFilter is moved to the end of the list to help with speed.
If the pool is already filtered out by any of the other filters, then
there is no reason to calculate the capacity factors for the pool.
This is why the CapacityFilter is at the end of the list.