[placement] Add db migration jobs
When running Placement on its own images, Placement is responsible for
migrating the DB. This is done via the placement-manage command.

We add 2 Jobs here - one for migrated the schema and one for doing any
necessary data migrations.

Both Jobs only run if they find "placement-manage" in the path and thus
will only do any work once we upgrade Placement to Yoga and its own
images.