[nova] Increase OpenstackNovaComputeIdle wait time
More often than not, the alert fires without real reason nowadays and
thus produces alert-fatigue. We now increase the time the condition has
to be true to make the alert fire to 60min. We won't get alerted early,
but we will still get notified of persisting problems.

If this proves to be too long, we will probably remove the alert in the
future. It's a very broad alert-condition and the cc3test suite should
provide similar information.