swift: roll daemonsets faster
This is based on my own experience driving deployments in a way that
does not make them a huge time sink. The most controversial part is
probably the `maxUnavailable: 2` on the swift-servers, but this is also
the most important in terms of reducing the time required to complete a
deployment.