[nova] Allow 16 GiB VMs on small HVs
Instead of allowing only 8 GiB VMs on the small HVs, we now allow 16
GiB, too, because this currently fits in nearly all regions. In the
regions it doesn't fit, we assume that most 16 GiB VMs will not be
migrated in the near future and thus not clog the small HVs.