[Cinder] Rollback volume stacking on datastores
This patch changes the default capacity weigher to spread volumes
out over datastores from stacking.  We have to do this short term to
help buy time for allowing volume snapshots on datastores.  Stacking
causes datastores to get full quickly.  When a datastore is full
snapshotting fails due to not having space to do the snapshot, which is
a full clone.