[cinder] Add support_group label to all alerts
We need to bump mariadb and memcached chart versions to get support for
setting the support_group label on the included alerts and we also have
to include the owner-info chart for getting support for setting
suport_group on the k8s objects' alerts.

Bumping mariadb to >= 0.6.0 needs an explicit password set for the
backup user.