[placement] Set logging to INFO
We didn't have logging for placement on INFO since we switched from
Nova-based images on Rocky to Placement's own images on Yoga, because we
forgot to switch over the logger definition. With fixing the logger
here, we get back INFO-level logging and by that global request ids back
in the log lines of all requests.