[Ironic] Limit authentication to initial connection
Prior this change, nginx would expire the url for all
requests. The problem here is that the javascript app
running in the browser will periodically do a POST
request with a session id and possibly commands
to the url, and the validation will break that.

To make matters worse, the app won't close the session
normally, keeping the shell on the server running
until it timeouts.

The session-id replaces the authenticated url as a secret,
so we do not need to validate the POST requests anymore.

In similar vein, validating the url for static resources
is superfluous, and we can pass those requests right
through.