[nova] use oslo.cache MemcacheClientPool for keystone_auth tokens
Nova has a high amount of open memcached connections in prod.
This might be related to keystonemiddleware not using oslo.cache
but an old in-house CacheClient implementation.

There is a good explanation of this issue when upstream enabled
oslo.cache in Zed by default [1].

Neutron has already enabled this in prod a long time ago [2].

[1] https://github.com/sapcc/keystonemiddleware/commit/788d3c4969e3446778496f3a9055f654602ae2c1
[2] https://github.com/sapcc/helm-charts/commit/0a0f884ebeb6f9e2e9d944bac89d1cbc3561f2de