[nova] Add pre-upgrade service version bump Job
Some time ago, Nova introduced a check in scheduler and conductor that
prohibits starting these services if there are compute services with a
too old version around. This version is read from the DB and is supposed
to be not older than 2 releases.

Since we want to do a big multi-version jump from Rocky to Xena, the
conductors will not come up with the Xena image because of the old
compute versions - but the new compute services also cannot come up to
set their newer versions, because they need the conductor for DB access.

To get out of this circular dependency, we run a k8s Job when we deploy
a "xena"-prefixed imageVersion that sets the versions of all compute
hosts to 54. This ensures a still supported version to let conductor
come up, but is not the version set by upgrade compute services. Having
an older version than Xena enables us to see if compute services did not
update their versions correctly and thus maybe didn't come up properly.

Updating the version needs to be done in all cells holding services.