[manila] disable determining replica sync state by transfer size
It is misleading. Response is hard to interpret:
'On last transfer there has been an amount of data replicated,
which was smaller than this limit'
It is merely an indicator if the users writing on the source share
have stopped writing. But this is something that the caller of
replica promote should already know or at least know to take this
into consideration.