[nova] Set initial allocation ratios
These settings have defaults we do not want - e.g. 16 x CPU overcommit
and 1.5 x RAM overcommit. They are only used to set defaults when a new
ComputeNode object is created and during an online data migration when
upgrading to Xena if they are set to None or 0. The allocation ratios
are then overwritten by the nova-compute service later on.