[nova] Shellinabox & Dell KMS Proxy in Nginx
This implements the shellinabox proxy in nginx
as lua scripts, and also solves proxying kvm
for dell hosts.

The existing custom solution in nova only proxies
to a single host, and cannot be changed easily
to multiple.

For the Dell KMS, it also needs to
proxy websockets, which is simply a configuration
option in nginx, but would otherwise need
to be implemented in the python proxy