memcached: Move yielded connections alert to info
This alert is not actionable, as it is at best a hint for possible
performance degradations - yielded connections are connections doing
multiple requests in batches and memcached interrupting to read from
that same connection to give other connections a chance to get processed
and not starve out.

We also make the threshold for the alert configurable as different
services might have different thresholds.