[nova] Set password_all_group_samples to 2
With ESXi ignoring an upper-case character at the start and a number at
the end of a password for counting the used and required character
classes, we need to use every class at least twice to make the password
fit the policy. Nova recently added code to make this possible, so we're
setting the password_all_group_samples setting here to enable it.