[sap-seeds] Fix cores-per-socket for HANA flavors
Nova's VMware driver uses the number given in `hw:cpu_cores` to set
the CoresPerSocket VMware setting for the VM[1]. While I think that's
probably a misunderstanding of what CoresPerSocket means by the
original author (description in [2], compare with "numCoresPerSocket"
in [3]), and should probably be fixed in the driver, let's do the
quick fix first and adjust the values for the flavors.

We thus use the NUMA node CPU number for the HANA flavors, which means
48 for most HANA flavors/BBs and 56 for the 12TiB flavors & BBs. This
aligns with the trait names CUSTOM_NUMASIZE_C48_M729 and
CUSTOM_NUMASIZE_C56_M1459 respectively.

Also, add the same "hw:cpu_cores" for all smaller HANA flavors,
because otherwise they too get smeared all over the sockets and who
wants to clean that mess up, really.

[1] https://github.com/sapcc/nova/blob/de8710b/nova/virt/vmwareapi/vmops.py#L468
[2] https://review.opendev.org/c/openstack/nova/+/187942
[3] https://vdc-repo.vmware.com/vmwb-repository/dcr-public/6b586ed2-655c-49d9-9029-bc416323cb22/fa0b429a-a695-4c11-b7d2-2cbc284049dc/doc/vim.vm.VirtualHardware.html