elektra: use new common/pgbackup chart
This bumps backup-tools from 0.6.5 to 0.7.0 and moves it into a separate
pod, so it can be upgraded in the future without having to restart the
postgres container. It also gets adjusted resource limits with a
significantly increased CPU limit, meaning that individual backups
should take less time to create in the future. The RAM limit is
significantly decreased, however, since the backup container never uses
more than around 150 MiB in practice.