[neutron] Flip network_owner_or_owner policy for performance
The check rule:owner is much faster than the check rule:network_owner,
as rule:network_owner needs to look up a network in neutron, which means
another DB query + dict buildup in python. rule:owner only looks up
fields that are already available on the port that the policy is checked
for.

The rule network_owner_or_owner is mainly used by get_ports, so
hopefully this increases the performance for port lists in Neutron for
ports in non-shared networks by users that are scoped to a project.

Co-Authored-By: Johannes Kulik <johannes.kulik@sap.com>