[neutron] Use prometheus to check if an agent is ready
The only constraint for a network is that is to be scheduled on any 2
network agents. Kubernetes does not know about a networks scheduling
and hence which agents can be absent at the same time without creating
an outage. As a consequence we can only safely reload one network agent
at a time. We also must make sure that all networks are synced to that
agent before rolling over a second.
In the past this was done querying the neutron-api, asking for agents
assgigned to the agent and then comparing that to the netns/bridges
locally. This put some load on the api as we have a few dozen network
agents in each region, each asking the api for their networks.
With the network-node-exporter in place, we can also ask
prometheus-openstack for that information, shifting that load from the
neutron-api to prometheus, which pretty much already has that data lying
around.

I hope this
1) reduces our dependency to a slow neutron-api leading to failing
   readyness probes,
and
2) reduces the load generated though dozens of agents asking for their
   assigned networks.