[neutron-network-agent] Use status file to check if an agent is ready
In the past this was done querying the neutron-api, asking for agents
assgigned to the agent and then comparing that to the netns/bridges
locally. This put some load on the api as we have a few dozen network
agents in each region, each asking the api for their networks.
With the network-node-exporter in place, we can also ask
prometheus-openstack for that information, shifting that load from the
neutron-api to prometheus, which pretty much already has that data lying
around.

I hope this
1) reduces our dependency to a slow neutron-api leading to failing
   readyness probes,
and
2) reduces the load generated though dozens of agents asking for their
   assigned networks.