[neutron] Use "host" variable for "vcenter" label
In 96cea43ef5 we changed all VCenterTemplate objects to include a
vcenter label. All other templates use the "host" variable to set the
"vcenter" label, except the nxv3-agent-configmap, which uses the
"vcenter_name" variable.

We change the configmap to use the "host" variable to make the label
consistent over all VCenterTemplates.