Ironic: Add setting to controll disk erase
This setting allows us to enable or disable the erasure of the disk
during the cleaning of a node. For now we want to disable this. Later we
can then enable it again. The other cleaning steps shouldn't cause major
issues but by enabling automated cleaning for some regions blocks we can
test if we find any other issues.