[octavia] Add loadbalancer id to "stuck in" alert
I like to have the id of everything handy, as most of our tooling works
well with uuids but takes a longer time with names. Therefore we now
have the loadbalancer id together with the loadbalancer name in the
alert text, same as we already do for our network alerts.