[PX] Fix pods being unschedulable during depoloyment
In regions with only 2 apods we cannot satisfy the scheduling constraint
while the old pods are still running because the anti affinity requires
the pods of one service and domain to be scheduled mutually
exclusive. That means we need to apods per domain. When 2 new pods appear
those would need additional 2 apods.

 Please enter the commit message for your changes. Lines starting