[PX] Sort the apods, this time for real
We were seeing rolling pods that had no actual change in the resource
definition other than the apod order being changed. While we thought
this was fixed in a08f6a2f1 it only sorted the list that was mapped to
the site. What still kept in an unstable order was the keys of $apods.
Let us now order the key names of $apods and access them in that order.