[nova] Set image default architecture for scheduler
The scheduler decides according to the image properties
on which hypervisor and in which mode to put the vm.

If hw_architecture isn't set in the image, the scheduler
might chose a host with a different architecture.
I.e. put the VM on an arm64 hypervisor (potentially emulated)

In order to avoid trouble should we ever introduce another
one, we can set the default value to reflect the default
assumption.