Add env-vars snippet for pyroscope
The snippet should take care of setting all the environment variables,
taking the defaults from utils, allowing an chart level override, but
also one on an application-level.

It requires a python user-customization file merge in sapcc/loci#15,
and additionally, the process cannot be run as root,
as a root process in a container is more restricted than a non-root user
one.

All settings from pyroscope can be set via the `PYRO_*` environment
variables.

Additionally, there is the `PYRO_APP_SUBSET` variable,
which is a regular expression, which needs to match the hostname,
for the agent to be come active.
The idea is, that you can thereby select a subset of the application
instances to be sampled for a potential A/B split.