[PX] Increase grace time for PXTransportAllRouterBgpDown
There was no grace time for this alerts as we deemed a situation in
which the router server is up but all transport router peers are gone as
very critical. What we did not think about if the container receives a
SIGTERM, bird in turn properly closes all BGP sessions orderly but the
exporter still exposes metrics. While we would not alert on no metrics,
we alert on metrics indicating sessions down. From SIGTERM to container
deleted it takes no longer than 90 secs, so increasing this threshold to
2min.