[cinder] Switch to utils chart's coordination ini section
We want to switch the coordination backend in the future and thus switch
to the ini section template for [coordination] from the utils chart - it
will make our switching easier.

For this, we have to bump the utils chart version to at least 0.7.0. In
the process, we also get a lot of updates to the proxysql handling.
Rendering the Cinder chart before and after the version bump shows only
differences in the proxysql image version.