[nova] Split out db with projected volumes plus minor cleanups
Instead of injecting the config for the databases in each config
use the default config dir logic and put place the config
under /etc/nova/nova.conf.d/db.conf, where all *.conf files
will be loaded.

Since the database config contains the db credentials, place
the file in a secret.

Use a projected volume to bind it all together.

Remove
- `policy.yaml` from nova-scheduler
-  iptables-restore completely
as they are not used