[sap-seeds] Extract common flags to snippets, set hypervisor to vmware
We want to be able to allow users to target another hypervisor (libvirt)
in the future, so for compatability reasons, the existing flavors
need to target vsphere.

We chose flavors over image attributes as it is more explicit,
and we only want allow boot-from-volume on KVM, while the existing
flavors have root-disks.

As a side-effect, it "breaks" forcing those non-baremetal flavor
to a baremetal host.