[openstack/utils] Snippet for kubernetes-entrypoint in initcontainer
Moving the logic to an initcontainer means we do not have to build
in the executable in the main image and can share that image among
services.

Also, a missing dependency won't get mixed up with "normal" program
failures.

The snippet could look used this:
```
      initContainers:
      {{- tuple . (dict "service" "service-mariadb;service-rabbitmq"
			"jobs" "db-migration") |
         include "utils.snippets.kubernetes_entrypoint_init_container" |
         indent 6 }}
```
(Except it needs to be on a single line, as the linter doesn't
understand multi-line gotemplates