[ironic] Randomize job name
If the job changes, it causes an error on validation
with three-way merge, despite the job being marked
as hook-delete-policy: before-hook-creation.
So, randomize the name to ensure there is no conflict.

Remove it as dependency from the api-deployment,
since that dependency is already solved
by helm itself (it won't apply the job, unless
the job has been executed successfully)