[nova] Switch nova-vspc version variable
Instead of using `vspc.version` we switch to using `imageVersionVspc`.
We need to do this, because we want to use the automatic version
transport of the shared pipeline to regularly get security updates in
the image. The automatic version transport needs the version to be
tracked in a variable adhering to a regex and being on the top level,
which `vspc.version` both does not satisfy.