[sap-seeds] Set disk for "forced" flavor to 1 GiB
The current Openstack-Seeder cannot deal with zero-sized disks because
it treats 0 as "empty" and prunes the key from the input flavor spec.

As a band-aid we set disk-size to 1 GiB for the KVM testing flavor,
but this needs to be reverted before the rollout of KVM since KVM
flavors shouldn't have ephemeral root disks at all.