[nova] Drop unused neutron references (imageVersion & ml2_vmware agent)
We moved most neutron logic out with the NSX-T agent in to the neutron
chart, so we can drop a lot of references to neutron images and
config-values.

Missing is the KVM hypervisor, but that needs to get addressed in a
different change to keep the change limited in scope.