[neutron] Topology aware scheduling and spreading + AZ update
Use topologySpreadConstraints to spread the api pods over the zones in a
best effort manner to improve the situation in the situation of a zone
is down (kubernetes may not rebalance, or cannot due to inbalance)

Use topology aware routing service-annotation to make use of the
distributed placement and prefer communicating to nodes in the same AZ

In regions with a single AZ, the annotations are not set.

Option to change the AZ affinity with global.topology_key to something
non-deprecated (i.e. topology.kubernetes.io/zone)

Add AZ affinity for neutron-nsxv3-agent deployment, so the agent
runs in the same AZ as the vcenter resides in