[nova] Disable uptime RPC call in hypervisor details
When fetching hypervisors from Nova's API, newer API versions (2.88+)
return an "uptime" field in the response. To fill this field, nova-api
makes an RPC call to each hypervisor. Since some of the drivers (namely
Ironic and VMware) do not implement the `get_host_uptime()` call and
thus end up raising a `NotImplementedError`. Making the RPC call for
every hypervisor is slow and since our hypervisors don't implement it
also clog the logs. Therefore, we disable the RPC call and always return
`None` for the field.