[nova] Switch to projected volume for nova-compute
Switching to projected volumes, we will work around a problem when one
of the containers in a pod crashes and a ConfigMap was edited since the
pod was started: the container would not be able to mount the old
ConfigMap anymore and not come up.
Additionally, we were told to get rid of mounts on top of mounts (e.g. a
file onto an emptyDir), because that can make things hang, too, iirc.