[Nova] Spawn consoles per cell
With Nova switching from the `consoleauth` service to storing tokens
inside the database, we need to spawn instances of the consoles per
cell, as the console services connect to one DB and rabbitmq only.
This also means, that we need to have different endpoints per cell, so
the right console service can be reached. To facilitate that, we change
the path in the URL for the console to include the cell.

Since the nova-compute pods the console config for the cell they run in,
we move the consoles' settings out of `nova.conf` and instead generate a
ConfigMap `nova-console` which contains one file per cell and
console-type. The console services mount only the one file they need
into their container. We use the `nova.conf.d` directory for not having
to explicitly set every config file on the command line for each
container.

Since cell2 is optional, we have multiple places that define a "$cells"
dict to iterate over. This is duplication, but I haven't found a good
way to move it into a central place.

shellinabox and Ironic hypervisors are hard-coded to `cell1`, because
Ironic does not live in cells but always lives in the cell1 DB, as the
nova-compute-ironic pods do not use different rabbitmq settings.

Each nova-compute pod (Ironic, VMware) have a list of console types
hard-coded in their template, as they only support these consoles and
don't need config for everything - especially does the VMware
nova-compute not need "shellinabox" console configuration for cell2, if
we don't have any Ironic in cell2. This enables us to not generate cell2
ConfigMap, Ingress and Service objects for some consoles in the future.